resolved telemetry conflict
